web scraping for getting profit for cards and new optimization rough draft for database, need to continue to refactor code

+++ Inventory/views.py
@@ -1,15 +1,34 @@
 import csv
 import io
+import re
 
+import requests
+from bs4 import BeautifulSoup
 from django.contrib import messages
+from django.db.models import F
 from django.shortcuts import render
+from django.views.generic import TemplateView
 from django_filters import FilterSet
 from django_filters.views import FilterView
 from django_tables2.views import SingleTableMixin
-from django.db.models import F
 
-from .models import Card, Sold
+from .models import Card, Sold, Bag, MarketEvaluation
 from .tables import CardTable, SoldTable
+
+
+class SetCodes:
+    SET_KEYS = {}
+    URL = "https://mtg.gamepedia.com/Template:List_of_Magic_sets"
+    page = requests.get(URL)
+    soup = BeautifulSoup(page.content, 'html.parser')
+    cards = soup.find('table', class_='wikitable')
+    row_columns = cards.find_all('tr')
+    for row in row_columns[1:]:
+        row_values = row.find_all('td')
+        set_name = re.sub("[\(\[].*?[\)\]]", "", row_values[1].text).rstrip()
+        set_abbrev = re.sub("[\(\[].*?[\)\]]", "", row_values[3].text).rstrip()
+        if set_abbrev != '':
+            SET_KEYS[set_abbrev] = set_name
 
 
 class CardFilter(FilterSet):
@@ -33,7 +52,8 @@ class CardTableView(SingleTableMixin, FilterView):
         # prompt is a context variable that can have different values      depending on their context
         prompt = {
             'order': 'name, edition, foil, quantity, bag',
-            'table': data
+            'table': data,
+            'set_keys': SetCodes.SET_KEYS
         }
         return render(request, template, prompt)
 
@@ -49,13 +69,31 @@ class CardTableView(SingleTableMixin, FilterView):
 
             io_string = io.StringIO(data_set)
             for column in csv.reader(io_string, delimiter=',', quotechar="|"):
+                name = (column[0].replace(';', ','))
+                edition = column[1]
+                foil = column[2]
+                bag = column[3]
+                quantity = column[4]
+                card_id = name + '_' + edition + '_' + foil
+                bag_id = card_id + '_' + bag
                 _, created = Card.objects.update_or_create(
-                    name=(column[0].replace(';', ',')),
-                    edition=column[1],
-                    foil=column[2],
-                    quantity=column[3],
-                    bag=column[4]
+                    id=card_id,
+                    name=name,
+                    edition=edition,
+                    foil=foil
                 )
+                bags = Bag.objects.filter(bag_id=bag_id)
+                if bags.count == 0:
+                    Bag.objects.update_or_create(
+                        bag_id=bag_id,
+                        id=card_id,
+                        bag_number=bag,
+                        quantity=quantity
+                    )
+                else:
+                    selected_bag = bags[0]
+                    selected_bag.quantity = F('quantity') + quantity
+                    selected_bag.save()
 
         context = {'table': Card.objects.all()}
         return render(request, template, context)
@@ -73,7 +111,6 @@ class SoldTableView(SingleTableMixin, FilterView):
         for sold in data:
             total_profit += sold.profit
 
-        print(total_profit)
         prompt = {
             'order': 'name, edition, foil, quantity, profit',
             'table': data,
@@ -96,27 +133,61 @@ class SoldTableView(SingleTableMixin, FilterView):
 
             io_string = io.StringIO(data_set)
             for column in csv.reader(io_string, delimiter=',', quotechar="|"):
-                card = Card.objects.filter(
-                    name=(column[0].replace(';', ',')),
-                    edition=column[1],
-                    foil=column[2],
-                    bag=column[4]
-                )
-                if card.count() > 0:
+                name = (column[0].replace(';', ','))
+                edition = column[1]
+                foil = column[2]
+                quantity = column[3]
+                card_id = name + '_' + edition + '_' + foil
+                card = Card.objects.filter(id=card_id)
+                card_bag = Bag.objects.filter(card_id=card_id)
+                if card_bag.count() > 0 and card.count() > 0:
+                    selected_card_bag = card_bag[0]
                     selected_card = card[0]
-                    if selected_card.quantity <= int(column[3]):
+                    if selected_card_bag.quantity <= int(quantity):
+                        selected_card_bag.delete()
                         selected_card.delete()
                     else:
-                        selected_card.quantity = F('quantity') - column[3]
-                        selected_card.save()
+                        selected_card_bag.quantity = F('quantity') - column[3]
+                        selected_card_bag.save()
                     _, created = Sold.objects.update_or_create(
                         id=selected_card.id,
-                        name=selected_card.name,
-                        edition=selected_card.edition,
-                        foil=selected_card.foil,
-                        quantity=column[3],
+                        quantity=quantity,
                         profit=column[5]
                     )
                     total_profit += column[5]
         context = {'table': Sold.objects.all(), 'total': total_profit}
         return render(request, template, context)
+
+
+class MarketAnalysis(TemplateView):
+    template_name = 'market.html'
+
+    def get(self, request):
+        template = 'market.html'
+        return render(request, template, {})
+
+    def post(self, request):
+        URL = "https://cardkingdom.com/purchasing/mtg_singles?filter%5Bipp%5D=100&filter%5Bsort%5D=name&filter%5Bnonfoil%5D=1&filter%5Bfoil%5D=1"
+        page = requests.get(URL)
+        soup = BeautifulSoup(page.content, 'html.parser')
+        cards = soup.find_all('div', class_='itemContentWrapper')
+        for card in cards:
+            card_name = card.find('span', class_='productDetailTitle').text
+            card_edition = card.find('div', class_='productDetailSet').find('a').text
+            card_foil = card.find('div', class_='foil') is not None
+            card_profit_dollar = float(card.find('span', class_='sellDollarAmount').text)
+            card_profit_cents = float(card.find('span', class_='sellCentsAmount').text) / 100
+            card_profit = card_profit_dollar + card_profit_cents
+        # pages = soup.find('ul', class_='pagination').find_all('a')
+        # page_list = list(range(1, int(pages[len(pages) - 1].text) + 1))
+        # for p in page_list:
+        #     page_url = 'https://cardkingdom.com/purchasing/mtg_singles?filter%%5Bipp%%5D=100&filter%%5Bsort%%5D=name&filter%%5Bnonfoil%%5D=1&filter%%5Bfoil%%5D=1&page=%s' %(p)
+        #     print(page_url)
+        #     next_page = requests.get(page_url)
+        #     soup = BeautifulSoup(next_page.content, 'html.parser')
+        #     cards = soup.find_all('div', class_='itemContentWrapper')
+        #     for card in cards:
+        #         card_name = card.find('span', class_='productDetailTitle').text
+        #         card_edition = card.find('div', class_='productDetailSet').find('a').text
+        template = 'market.html'
+        return render(request, template, {})
